Adds Initial HTTP Client
Adds an initial HTTP Client for interfacing with the github graphql api.

The original "clients" have been renamed as "repositories" (i.e.
HttpClient --> GithubRepo and LocalClient --> LocalRepo). The repository
will encapsulate the work to orchestrate retrieving all of the data we
need.

The new `GithubClient` will be the contact point between our app and the
Github api, handling the creation and parsing of requests/responses,
authentication, etc.